+++ docs/docs/changelog.md
@@ -15,6 +15,9 @@ Types of changes:
 - `Security` in case of vulnerabilities.
 
 ## [Unreleased] - {ISO8601 date}
+### Added
+- Added generic read_file function
+- Added generic read_dir function
 
 ## [0.8.0] - 2020-04-26
 ### Added